Created the macro "table_pending_answer" to hide the details of obtaining the next answer for consumption, which, in the case of a properly subsumed subgoal, may require an answer identification step.
Second, removed -- again -- (most) "xtemp1" references.
(Argh!  How does this crap keep creeping back into the system?!?)

These were sprinkled over several files and removed by simply creating
a variable local to each macro in which a temp was needed.

The "need" for this change was precipitated due to the inclusion of an
"xtemp1" declaration within one of the code blocks which
table_pending_answer replaced.


git-svn-id: http://svn.code.sf.net/p/xsb/src/trunk@2185 756869e0-9e47-4bc5-b0a8-847305dc5e1c