I've committed a version of CDF.  Its primarily for Luis to work with, but I figured I'd let everyone know what has been done.
0) This version is built on top of the initial pass at relative paths
in the ocmponent system that Luis put in last week.

1) The CDF loads up the cdf component and the cdfpt components upon
(re-) initialization through the component system.

2) Structures have been added as a primitive type -- which I think
will work where the old clause primitive type used to work.

3) I've started moving flags to cdf_flags/2.  In particular, I've made
the consistency flags controllable this way.  Currently the default is
to do classification in bulk upon component load, to check types in
newExtTerm, and to have the implication check turned off.


git-svn-id: http://svn.code.sf.net/p/xsb/src/trunk@3422 756869e0-9e47-4bc5-b0a8-847305dc5e1c