All:
I've updated CDF with a number of changes.  I'm pretty sure they wont
cause problems... but ... now that we're done with David's compiler
and system changes we need something to keep us busy :-)
We've talked about most of these changes.

1) I added coversAttr, with the different inheritance needed for time
intervals and spatial reasoning.  I also added concrete components for
which the inheritance of terms is not product class inheritance.  Both
cdfpt and cdftm are concrete classes.

2) I added the package cdftm that allows comparisons of James Allen's
intervals.  This is an XSB proprietary package

3) I made some changes to the theorem prover, and have started to
incorporate it into the documentation.  It is pretty much ready to
use.

4) I moved indices for extensional facts from cdf_init_cdf to
cdf_configure.  I do need to add a facility to read index declarations
from a user's xsbrc file, but will add that later.

5) I fixed the handling both of intensional rules, and dynamic rules
that can support intensional rules.  This makes it easier to maintain
large packages of rules via the components system.

6) As there are a lot of new features in CDF I've begun updating the
documentation.  Given the other distractions, it will probably be
several weeks before this is done.


git-svn-id: http://svn.code.sf.net/p/xsb/src/trunk@3772 756869e0-9e47-4bc5-b0a8-847305dc5e1c