Added new mutex, MUTEX_DELAY to synchronize management of de-s dl-s and pnde-s.  For now, memory management of these data structures uses common global variables, even in private tables, hence the need for a mutex.
My placing of the mutexes within the code is still a little
uncertain.  On the one hand, I've over approximated a little so
that (I think) any of the memory management routines are covered.
I've also covered simplification routines (which do not directly
call the mem-management routines), though I'm not entirely
certain that the mutexes are needed there with shared tables and
our algorithm for shared concurrent tables.  More insight on
mutexes and simplification will undoubtedly come with paper
writing.

In addition, we should think in the future about managing de-s,
dl-s and pnde's for private tables using thread-specific memory
management routines.


git-svn-id: http://svn.code.sf.net/p/xsb/src/trunk@4557 756869e0-9e47-4bc5-b0a8-847305dc5e1c