1. Changed the XSB tokenizer to be more compatible with the Prolog standard.  This involved adding escape characters when processing strings -- single quoted, double quoted and 0-radix, as in 0'\n.  The escape characters are those in the standard (taken from C).  I did not add \' as a single quote within a quoted string, since it breaks the form: '/\', which appears several times in the XSB system code.  (This could, of course, be changed.)  Backslashes that are not followed by a legal escape indicator are accepted simply as backslashes.  (I suspect this is not what the standard expects, but it will break fewer programs.)  I added the new forms for binary (e.g., 0b101), octal (e.g., 0o731), and hex (e.g., 0xffef) that the standard requires.  I left in the non-0 radix notation (after fixing a few bugs in it) which is not in the standard, so we still process e.g., 8'713 as an octal integer.  I extended the allowable radix up to 36 (which was partially there already in a buggy form; I used A-Z for the digits 10-35.)  This is probably not of much use, but what the hey....
The addition of the escape changed one of the tests in the testsuite.
Update tests from cvs to get the new correct answers.

2. Improved assert_code_to_buff so that it will run out of registers
less often when asserting GROUND structures.  The old version had a
limit of approximately 255 for right-branching when the left-elements
were non-constant.  E.g. arbitrary lists of constants were handled, but
lists of structures were limited to c. 255.  I've changed it so the
limit is now on left-branching instead of right-branching.  Since
right-branching is much more common in Prolog programs, this should
handle more structures that show up in Prolog programs.  (No claim is
made for Flora programs :-).


git-svn-id: http://svn.code.sf.net/p/xsb/src/trunk@4578 756869e0-9e47-4bc5-b0a8-847305dc5e1c