Fixed problems with throwing memory errors from mem_xxxoc(), at least in sequential engine.  When memory cannot be malloced, a resource error is thrown.  Since throwing errors mallocs memory in itself, mem_xxxocs used by a throw are changed into mem_xxxoc_nochecks and an xsb_exit is called if these functions return NULL pointers.  Care is taken so that when a resource error is thrown, space for its PSCs has been pre-allocated, and varstrings are used for the actual message strings (varstring exits too, if there is not enough memory).
In MT engine, a memory error just exits -- this is because I need
a CTXT to throw an error.  Right now, I'm not sure whether its a
good idea or not to put CTXTs into mem_xxxoc or not.

Not that some of the biggest memory allocations, such as tcp
stack, local-heap stack use straight reallocs and use their own
error handling.

In general, we should be handling most memory errors in a
reasonable, or at least consistent, manner.


git-svn-id: http://svn.code.sf.net/p/xsb/src/trunk@4603 756869e0-9e47-4bc5-b0a8-847305dc5e1c