A number of changes in this update.
1) there are changes to allow transitive subgoal abolishes for tables
with conditional answers.  If a subgoal S is abolished, then all
subgoals that (transitively) depend on S, positively or negatively,
will be abolished by default in order to prevent core dumps when
examining the residual program.  Only backward dependencies are
abolished -- forward dependencies do not cause a problem.

The transitive actions of the abolish can be turned off either by
a new xsb_flag or by calling a new predicate
abolish_table_call/2.

While testing these changes I discovered and fixed an obscure bug --
the table abolish CPS checks could not find the trie nodes when a
hash_handle choice point was on the choice point stack.  The result
was that not all tables with choice points were found when the CPS was
traversed.

2) There are similar changes to allow transitive predicate
abolishes for tables with conditional answers.  This is analogous
to the subgoal case above, but while doing this I discovered
another obscure bug.  It turns out that freeing a subgoal
frame (Free_Producer_SF()) changed the subgoal list that hung off
of the TIF.  This is the right action if space is reclaimed
immediately for a tabled predicate or subgoal, but this action should not be
done during garbage collection as the value of the subgoal list
off of the tif is for an entirely new set of tabled subgoals --
not the ones being reclaimed.

Both of these changes were a good excuse for (or required) refactoring
the table abolish and gc code.

Additions were made to the sequential and multi-threaded test suite to
test out the new functionality.

This update also includes some changes to interned tries that I
started in order to allow shared interned tries in the mt engine.
These aren't finished, but the interned trie code is in a consistent
state, so the update won't hurt anything.  They also included some
consistent but unfinished changes in alias handling.


git-svn-id: http://svn.code.sf.net/p/xsb/src/trunk@5303 756869e0-9e47-4bc5-b0a8-847305dc5e1c