Added --with-bits32 to allow 32-bit compilation on 64-bit platforms.
Changes for --with-bits32 and --with-bits64 so that these options
automatically extend the tag, and bits64 automatically uses fast floats.

Improved error checking for mispelled options that aren't handled by
configure.

Added consistency check to prevent with-bits32 and with-bits64.

Works on Mac.  Will test on linuxes and if it works there, I'll update
configure.


git-svn-id: http://svn.code.sf.net/p/xsb/src/trunk@5477 756869e0-9e47-4bc5-b0a8-847305dc5e1c