Two sets of changes.
First, was to add const declaration modifiers to external ctop and
ptoc routines.  Apparently some C++ compilers can benefit from this
modifier, and it should help the Parma polyhedra group and others who
call XSB from C++.

The second change supports call_cleanup, and fixes a minor bug in the
implementation of catch (see the emails on the XSB group).  This
update does not handle cutting over call cleanup -- I'll try to get
that in soon.


git-svn-id: http://svn.code.sf.net/p/xsb/src/trunk@5591 756869e0-9e47-4bc5-b0a8-847305dc5e1c