Fixed a bug in the XSB command loop where XSB exited after thorwing an exception that was caught at the top-level.
A few months ago, I made a change so that the top-level catch
resets the current input stream to whatever it had been before
the catch.  This was necessary to properly recover from errors
thown out of predicates that had reset standard input.  However,
it turned out that if XSB was called with the -e option, the
current input had not been reset after executing the command-line
goal.  As a result, upon catching the error the current input
stream was reset to the command-line string, the eof was reached
and XSB exited.  So the change now resets the current input
stream to whatever it was *before* the command line goal had been
executed -- i.e. current input (except possibly in some weird
cases).


git-svn-id: http://svn.code.sf.net/p/xsb/src/trunk@6156 756869e0-9e47-4bc5-b0a8-847305dc5e1c