Some mostly simple changes in memory management involving mem_calloc. mem_calloc was changed to not return NULL if memory is not available but to throw an error.  This was generally a good idea.  It made some error handling by callers of mem_calloc obsolete, so most of these changes are simply to delete that unnecessary code.
A bigger change is to try to allow findall to recover from a memory
overflow.  If findall fails to get memory, an error had been thrown,
but none of now-unused space was released.  So these changes release
that space.  (So a new mem_calloc_nocheck function was added similar
to the other _nocheck versions, so the caller can do more memory
management.)  So now findall releases all buffers it holds when it
can't get more memory.  I had hoped this would allow graceful
recovery, but it seems that even if we release all the space back to
the system, memory is now fragmented and future requests for large
amounts of space (as for expanding the heap) now fail.  So the
recovery is still not that good, but it's better than before.


git-svn-id: http://svn.code.sf.net/p/xsb/src/trunk@6173 756869e0-9e47-4bc5-b0a8-847305dc5e1c