Changes were mostly for documentation and code refactoring.
However, also fixed a latent bug in simplification where
was_simplifiable was not properly checking negative delay elements
corresponding to subsumed subgoal frames.  This led to the possibility
of adding un-needed delay elements that might not have been simplified
away.  Surprisingly enough, however, none of the tests in the test
suite checked for this.


git-svn-id: http://svn.code.sf.net/p/xsb/src/trunk@6205 756869e0-9e47-4bc5-b0a8-847305dc5e1c