Extended (almost) all the marking routines for tabled predicates so that subgoals in the heap delay list are checked.  This rather obscure bug caused a crash in a Flora program for defeasible logic.  I'll try to finish that part tomorrow.
The problem was that space was reclaimed for a table T that had a
delay element in the heap pointing to it.  When the answer with
the delay element was added, the engine tried to adjust the
backpointers for T and crashed.

Also made some minor ajustments for efficency to some of the
marking functions.  And I changed abolish_table_info to the
clearer abolish_all_tables.


git-svn-id: http://svn.code.sf.net/p/xsb/src/trunk@6227 756869e0-9e47-4bc5-b0a8-847305dc5e1c