This change affects excecution of table p/n as variant/subsuptive.
Previously, a test such as that below would have dumped core when
changing from variant to subsuption, so a couple of days ago I fixed
it so that it threw a permission error.  However, in thinking about
it, it may be safe to change the evaluation method even if a predicate
already has tables.  The idea is that new tables will get the method
from the psc or tip, while incomplete tables will get the method from
the subgoal frame.  Im not sure if this will work in general, but it
does work for the following test.  So for now, Im allowing it, but
printing out a warning,	    so I can fix things when they go wrong.

This will hopefully give us a measure of adaptivity for	  call
subsuption/variance, if a profiler indicates that the evaluation
method should change during a session.


test:- p(X),fail.
test:- p(1),fail.
test:- (table p/1 as subsumptive),fail.
test:- p(2),fail.
test:- (table p/1 as variant),fail.
test:- p(3).

:- table p/1.
p(X):- writeln(executing_clause_body(p(X))).


git-svn-id: http://svn.code.sf.net/p/xsb/src/trunk@6578 756869e0-9e47-4bc5-b0a8-847305dc5e1c