Implemented (partially) call abstraction for subsumptive tabling. E.g. :- use_subsumptive_tabling p(+,-,-), p(-,-,-).
causes a call to p/3 that has a mode more specific than (+,-,-) to be
abstracted to (+,-,-), which is the call that is made.  The second
declaration here for the same predicate, p(-,-,-), allows a most
general call to be made.  When a call is made, the goals are checked
in order to find the first one that has the same or more general mode
than the call, and that abstraction (if any) is done.  If none is
found, an error is thrown.

I've found this functionality very useful. E.g., when the goal defines
a database call (using an SQL query), this will enforce the correct
mode required by the SQL query automatically.  It can also be used to
implement more complete indexing.

I have not fully integerated it into XSB (use_subsumptive_tabling is
the only directive that supports this syntax.)  To fully implement it
would be more compex, requiring runtime transformations as well.  I
have not documented it in the manual, because of these limitations.  I
am committing the changes since I find them useful.  If others do too,
(and/or I get more time and interest), I will more fully integrate it,
and document it.


git-svn-id: http://svn.code.sf.net/p/xsb/src/trunk@6785 756869e0-9e47-4bc5-b0a8-847305dc5e1c