Bug fix.
Fixed copy/paste errors in nunify_with_num, nunify_with_nil, and
nunify_with_con.

The error lead to core dumps in the predicate

qr1(X):- when(nonvar(X),X \= 2),
        r(Y),
        X is Y + 1.

as the specialized instruction addintfastuni was called, which
called nunify_with_num() with its first parameter as op3 rather
than op1 as in all other uses of this macro.  This hit a bug in
the maro which had op1 in the attv case, rather than the macro
parameter OP1.

This bug only arose when attvs were used in arithmetic
expressions that used the instruction addintfastuni.


git-svn-id: http://svn.code.sf.net/p/xsb/src/trunk@6818 756869e0-9e47-4bc5-b0a8-847305dc5e1c