Changes	to add stack expansion so that arbitrary numbers of variables are allowed in interned, asserted tries and tables.
This uncovered a bug.

The problem arose when a conditional answer is interned in the
answer trie and that conditional answer contains variables both
in the answer head and the delay list, and there are variables in
the answer head that are not in the delay list. What happened was
that when the variables were interned in delay_trie_chk_insert,
the mini-trail that the trie routines use for standardizing
variables was reset.  This meant that the variables in the answer
head were not being properly untrailed and, due to the particular
way that variables are standardized, variables were bound to each
other that shouldn't have been, and in this case a circular
dereferencing chain arose, so that XSB hung.  I have no idea why
the bug was uncovered just now.


git-svn-id: http://svn.code.sf.net/p/xsb/src/trunk@7065 756869e0-9e47-4bc5-b0a8-847305dc5e1c