Several fixes: 1. interprolog_callback on MT had a minor issue.
2. a couple of casts to avoid warnings

3. Most of the changes are due to simplified handling of short and
boxed integers.  A bug was submitted that showed an error is handling
of boxed integers.  To fix that, I added an isointeger test macro to
test whether a prolog_term is an integer OR a boxed integer.  (We had
done this for boxed floats, but not integers.)  Anyway, using this
simplified much code, so I went ahead and made many changes, finding a
and fixing number of minor bugs in handling large integers.


git-svn-id: http://svn.code.sf.net/p/xsb/src/trunk@7078 756869e0-9e47-4bc5-b0a8-847305dc5e1c