Changed	some heap expansion thresholds to use OVERFLOW_MARGIN rather than the hard-wired	    2000.  OVERFLOW_MARGIN is in fact a settable flag.
The idea is that by setting OVERFLOW_MARGIN a user may be able to
affect to some extent the amount of heap gc that occurs.  If after
heap gc	  there are still not OVERFLOW_MARGIN cells between the heap and
local stack, expansion will occur.  Thus by setting OVERFLOW_MARGIN
high, we'll expand more	    aggressively.

The fly	    in the ointment is dynamic code.  The test_heap	instruction
contains an argument representing an overflow margin at the entry to
clauses for a predicate.  For static code this is always 2000, but for
dynamic code it is function of the size of the clauses for a given
predicate.  I'd been hoping that OVERFLOW_MARGIN could replace the
test-heap argument, but it looks like it cant -- not without having a
separate test_help instruction for static and dynamic code.

So OVERFLOW_MARGIN will *sometimes* affect heap	      gc.


git-svn-id: http://svn.code.sf.net/p/xsb/src/trunk@7422 756869e0-9e47-4bc5-b0a8-847305dc5e1c