Changed some index defaults,
and added a cdf_flag (implicit_isa) to turn off isa-check for product
data types.  With this on, an unindexed call to isa_ext is generated,
which can be very expensive, and unless an application needs full
(well, fuller) isa through product-types, one would rather avoid this
expense.  This flag allows that.


git-svn-id: http://svn.code.sf.net/p/xsb/src/trunk@7548 756869e0-9e47-4bc5-b0a8-847305dc5e1c